Reduce number of CI builds chefspec results in
Would it be possible to only test against the latest 0.9.x and 0.10.x?

chefspec run takes on average 12 minutes. Right now it results in over 20 builds per push. Simple math suggests that is 4 hours of run time. Even though we now have 10 workers on travis-ci.org, it is a big difficult to justify testing against 0.9.12 vs 0.9.18 if it leads to run times like this.

Thank you.