Don't include the deprecations module by default.
In the next release of ChefSpec, the deprecations module will not be required
by default. Because the module is a monkey-patch on top of existing monkey
patches, it weighs a significant overhead loading it on each run. It also
provides no benefit to new users of ChefSpec or users who have already
converted their code to use the ChefSpec 3 matchers.

The deprecations module still exists, but it will no longer be included by
default. Users still wishing to switch from ChefSpec 2.x can manually require
the module, as outlined in the README.

This module will be completely removed in ChefSpec 4.