Add ChefSpec::Cacher module for faster specs
The cacher module allows for ultra-fast tests by caching the results of a
Chef Client Run in memory across an example group. In testing, this can
reduce the total testing time by a factor of 10x. This strategy is _not_ the
default behavior, because it has implications surrounding stubbing and is
_not_ threadsafe!

- Refs: #275
- Refs: http://dracoater.blogspot.com/2013/12/testing-chef-cookbooks-part-25-speeding.html
- Credits: @DracoAter