Restart work towards a working minimal example of chef-search
This is neither minimal nor is it working. I can't seem to figure out how to mock a node with working fqdn (that is not fauxhai.local).

Interestingly the second way of mocking the node by converging with a recipe gives the 'bar' attribute as defined in the recipe. So that is working nice.