Fix setting cookbooks folder from spec on windows sets cookbooks to C:\
An issue was found on Windows systems where attempting to infer the
cookbooks directory from the calling spec was setting the inferred
directory to /, which means the root of the current drive on Windows.
This was caused because the : in the path was being inferred as the
start of the caller's line number.