Filter all resources that lack a `source_line` attribute
Due to a bug in Chef, when resources are cloned (specifically the `service`
resource), both the clone *and the original resource* lose all metadata
information including their source line. Chef claims this is not a bug.

As a result, ChefSpec cannot reliably check if a resource has been covered
when it was cloned. This commit excludes all resources that lack the proper
metadata to determine if they have been covered.