Add environment_path support
In the case that you're testing a cookbook as part of a larger chef repo
ideally, you'd be able to test for the existence of environment
attributes as well.

This commit adds support for an environment_path configuration variable
to parallel the role_path.

It seems unlikely that this will have a change in behavior, unless
someone has an environments/_default{.json,.rb} file or has explicitly
set `node.chef_environment = production` in their chefspec test. In the
latter case, this actually provides better testing because the node
more-closely matches production.