Add have_attribute matchers to Chef::Resource
This adds has_(attribute)? methods to Chef::Resource and its
descendants, enabling testers to write blocks like the following in the
context of a chef resource as the subject:

it { is_expected.to have_performed_action(:create) }

(for a user resource)
it { is_expected.to have_members([user]) }

(for a file resource)
it { is_expected.to have_mode('0640') }

and so on and so forth. I think this could allow users to better follow
the guidelines at betterspecs.org, and also has great benefits with
--format documentation output.