fix for chef-13 removal of resource cloning
i'm not positive this ever quite worked correctly, and there's probably
chef-12 bugs that this closes.

the old code assumed one resource would match the type+name and that it
would also match the action.  if there were two resources it would pick
the first (probably wrong based on how the chef resource collection
works) and then the matcher code would try to match the action.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>