Defang the Cookbook GemInstaller
The number of Cookbooks declaring gem dependencies in their metadata is
starting to grow. The implementation of the Cookbook GemInstaller uses
'bundle install' to fetch the required gems.
In a testing environment such behavior should be deactivated, to avoid
modification of the test machine system.
Also, in most cases ChefSpec won't need the gem itself because the code
using it is not called - resource's actions code is not evaluated.

This commit monkey-patch the Chef Cookbook GemInstaller to not perform
the `bundle install` command, but instead warn the user about missing
gems.

Fix #871