Make Berksfile path configurable
- In berkshelf.rb, replace 'from_file' by 'from_options'.
This allows to specify the path on Berksfile by configuring the
'berksfile' entry in berkshelf_options. It is useful in particular
if we manage several cookbooks in a single Chef repository (so with a
common Berksfile).
By the way, from a programmatic point of view, it removes a magic value
(the default name of the Berksfile) that is already given by Berkshelf
module.