Update chefspec for after_resource
This changes the chefspec API to target the "current_value" which
is short for "load_current_value" aka both the current_resource and
the after_resource objects.

This seems like the best API since the current_resource and the
after_resource shouldn't really be treated all that differently
(it is also not clear to me that anyone will really care at all
about chefspec'ing the after_resource -- chefspec has nothing to do
with testing the data collector output from the resource report).

So, I haven't included the ability to stub the current_resource
or after_resource separately, and made the current_value
stubbing backwards compatible with the current_resource stubbing.

At some point if it comes up we could consider splitting out
explicit stubbing of the current_resource + after_resource, but
that would probably be an API break when the current_resource
stubbing changes (again, though, I can't see how we need this
in chefspec and what people would be testing).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>