Add support for evaluating lazy attributes in templates
Previously if you had a template with `variables` that used the `lazy`
keyword, previously it was not possible to validate that the blocks
were evaluated in the template. The rendered string would contain an
`<Chef::DelayedEvaluator>` string.

This commit renders the variables to ensure the template renderer
displays the lazily-evaluated values.