Pin rspec-expectations due to private API change
In this change, we are introducing a temporary version pin
of the rspec-expectations gem to version v3.12.3

As part of the v3.12.4/v3.13.0 releases, the private API class
RSpec::Matchers::ExpectedsForMultipleDiffs was renamed to
RSpec::Matchers::MultiMatcherDiff [1]. chefspec does make use of this
class in one instance.

A future change should be introduced to properly address this
class dependency. Possible options include replicating a minimal
replacement, or porting the class to equivalent custom matcher
with the chefspec codebase.

[1]: https://github.com/rspec/rspec-expectations/commit/d0bb212

Signed-off-by: Joshua Riojas <joshuariojas.psu@gmail.com>